Refactor audio manager class in preparation of supporting macOS
